+++ src/integrations/integrations.test.ts
@@ -5,6 +5,7 @@ import {
     GraphQLSchema,
     GraphQLObjectType,
     GraphQLString,
+    GraphQLInt,
     GraphQLError,
     introspectionQuery,
     BREAK,
@@ -30,7 +31,7 @@ const QueryType = new GraphQLObjectType({
         testContext: {
             type: GraphQLString,
             resolve(_, args, context) {
-                return context;
+                return context.testField;
             },
         },
         testRootValue: {
@@ -44,6 +45,26 @@ const QueryType = new GraphQLObjectType({
             args: { echo: { type: GraphQLString } },
             resolve(root, { echo }) {
                 return `hello ${echo}`;
+            },
+        },
+        testBatchIndex: {
+            type: GraphQLInt,
+            resolve(root, _, context) {
+                if (context) {
+                    return context.apolloBatchIndex;
+                } else {
+                    return null;
+                }
+            },
+        },
+        testBatchSize: {
+            type: GraphQLInt,
+            resolve(root, _, context) {
+                if (context) {
+                    return context.apolloBatchSize;
+                } else {
+                    return null;
+                }
             },
         },
         testError: {
@@ -332,6 +353,61 @@ export default (createApp: CreateAppFunc, destroyApp?: DestroyAppFunc) => {
           });
       });
 
+       it('puts batch index in context', () => {
+          app = createApp();
+          const expected = [
+              {
+                  data: {
+                      testBatchIndex: 0,
+                      testBatchSize: 3,
+                  },
+              },
+              {
+                  data: {
+                      testString : 'it works',
+                  },
+              },
+              {
+                  data: {
+                      testBatchIndex: 2,
+                  },
+              },
+          ];
+          const req = request(app)
+              .post('/graphql')
+              .send([{
+                  query: `query test1 { testBatchIndex, testBatchSize }`,
+              }, {
+                  query: `query test2 { testString }`,
+              }, {
+                  query: `query test3 { testBatchIndex }`,
+              }]);
+          return req.then((res) => {
+              expect(res.status).to.equal(200);
+              return expect(res.body).to.deep.equal(expected);
+          });
+      });
+
+
+       it('does not put batch index in context', () => {
+          app = createApp();
+          const expected = {
+              data: {
+                  testBatchIndex: null,
+                  testBatchSize: null,
+              },
+          };
+          const req = request(app)
+              .post('/graphql')
+              .send({
+                  query: `query test1 { testBatchIndex, testBatchSize }`,
+              });
+          return req.then((res) => {
+              expect(res.status).to.equal(200);
+              return expect(res.body).to.deep.equal(expected);
+          });
+      });
+
       it('can handle a request with a mutation', () => {
           app = createApp();
           const expected = {
@@ -373,7 +449,7 @@ export default (createApp: CreateAppFunc, destroyApp?: DestroyAppFunc) => {
           const expected = 'context works';
           app = createApp({apolloOptions: {
               schema: Schema,
-              context: expected,
+              context: {testField: expected},
           }});
           const req = request(app)
               .post('/graphql')